+++ plugins/voxelman/gui/plugin.d
@@ -48,6 +48,7 @@ final class GuiPlugin : IPlugin
 private:
 	EventDispatcherPlugin evDispatcher;
 	ConfigOption resolution;
+	ConfigOption maxFpsOpt;
 
 public:
 	IWindow window;
@@ -61,17 +62,20 @@ public:
 	{
 		auto config = resmanRegistry.getResourceManager!ConfigManager;
 		resolution = config.registerOption!(uint[])("resolution", [1280, 720]);
+		maxFpsOpt = config.registerOption!uint("max_fps", true);
 	}
 
 	override void preInit()
 	{
 		initLibs();
-		fpsHelper.limitFps = false;
 
 		window = new GlfwWindow();
 		window.init(uvec2(resolution.get!(uint[])), "Voxelman client");
 		renderer = new OglRenderer(window);
 		igState.init((cast(GlfwWindow)window).handle);
+
+		fpsHelper.maxFps = maxFpsOpt.get!uint;
+		if (fpsHelper.maxFps == 0) fpsHelper.limitFps = false;
 
 		// Bind events
 		window.windowResized.connect(&windowResized);