+++ plugins/voxelman/world/plugininfo.d
@@ -1,0 +1,6 @@
+module voxelman.world.plugininfo;
+enum id = "voxelman.world";
+enum semver = "0.6.0";
+enum deps = [];
+enum clientdeps = [];
+enum serverdeps = [];